implement fetching data from repository to the service

+++ module-service/src/main/java/com/mjc/school/service/NewsService.java
@@ -5,8 +5,8 @@ package com.mjc.school.service;
 import java.util.List;
 
 public interface NewsService {
-    boolean deleteNewsEntry(int id);
-    NewsDTO getNewsById(int id);
+    boolean deleteNewsEntry(long id);
+    NewsDTO getNewsById(long id);
     int createNewsEntry(NewsDTO news);
     List<NewsDTO> getAllNews();
     boolean updateNewsEntry(NewsDTO news);
